@@ -1,3 +1,2 @@
-from flask import Blueprint
-
-blueprint = Blueprint('data_manager_blueprint', __name__, template_folder='templates')
+"""This file and its contents are licensed under the Apache License 2.0. Please see the included NOTICE for copyright information and LICENSE for a copy of the license.
+"""
