@@ -1,2 +1,2 @@
-from .api import init_app
-from .model import LabelStudioMLBase
+"""This file and its contents are licensed under the Apache License 2.0. Please see the included NOTICE for copyright information and LICENSE for a copy of the license.
+"""
